Move helper traits to new pkg plasm-support (#268)
* initial

* downgrade log package

* fix formatter

* plasm version updates to 1.7.0
