+++ frame/dapps-staking/Cargo.toml
@@ -27,8 +27,9 @@ pallet-contracts = { version = "2.0.0", default_features = false }
 pallet-staking = { version = "2.0.0", default_features = false }
 sp-staking = { version = "2.0.0", default_features = false }
 
-pallet-contract-operator = { path = "../operator", version = "1.4.0", default_features = false }
-pallet-plasm-rewards = { path = "../plasm-rewards", version = "1.4.0", default_features = false }
+pallet-contract-operator = { path = "../operator", version = "1.7.0", default_features = false }
+pallet-plasm-rewards = { path = "../plasm-rewards", version = "1.7.0", default_features = false }
+pallet-plasm-support = { path = "../plasm-support", version = "1.7.0", default_features = false }
 
 [dev-dependencies]
 wabt = "0.10.0"
@@ -57,5 +58,6 @@ std = [
     "pallet-staking/std",
     "pallet-contract-operator/std",
     "pallet-plasm-rewards/std",
+    "pallet-plasm-support/std",
     "sp-staking/std",
 ]
