+++ frame/dapps-staking/src/lib.rs
@@ -16,11 +16,11 @@ use frame_support::{
     IterableStorageDoubleMap, StorageMap, StorageValue,
 };
 use frame_system::{self as system, ensure_signed};
-use pallet_contract_operator::ContractFinder;
 use pallet_plasm_rewards::{
     traits::{ComputeEraWithParam, EraFinder, ForDappsEraRewardFinder, HistoryDepthFinder},
     EraIndex, Releases,
 };
+use pallet_plasm_support::ContractFinder;
 pub use pallet_staking::{Forcing, RewardDestination};
 use sp_runtime::{
     traits::{AtLeast32BitUnsigned, CheckedSub, Saturating, StaticLookup, Zero},
