+++ frame/operator/Cargo.toml
@@ -21,6 +21,7 @@ frame-support = { version = "2.0.0", default_features = false }
 pallet-timestamp = { version = "2.0.0", default_features = false }
 frame-system = { version = "2.0.0", default_features = false }
 pallet-contracts = { version = "2.0.0", default_features = false }
+pallet-plasm-support = { path = "../plasm-support", version = "1.7.0", default_features = false }
 
 [dev-dependencies]
 wabt = "0.10.0"
@@ -43,4 +44,5 @@ std = [
     "pallet-timestamp/std",
     "frame-system/std",
     "pallet-contracts/std",
+    "pallet-plasm-support/std",
 ]
