+++ frame/operator/src/lib.rs
@@ -4,10 +4,7 @@ use frame_support::dispatch::UnfilteredDispatchable;
 use frame_support::{decl_event, decl_module, decl_storage, Parameter};
 use frame_system::{ensure_signed, RawOrigin};
 use pallet_contracts::{BalanceOf, CodeHash, ContractAddressFor, Gas};
-use sp_runtime::{
-    traits::{MaybeDisplay, MaybeSerialize, Member},
-    DispatchError,
-};
+use sp_runtime::traits::{MaybeDisplay, MaybeSerialize, Member};
 use sp_std::prelude::*;
 
 pub mod parameters;
@@ -16,50 +13,7 @@ mod tests;
 
 use crate::parameters::Verifiable;
 
-pub trait ContractFinder<AccountId, Parameter> {
-    fn is_exists_contract(contract_id: &AccountId) -> bool;
-    fn operator(contract_id: &AccountId) -> Option<AccountId>;
-    fn parameters(contract_id: &AccountId) -> Option<Parameter>;
-}
-
-pub trait OperatorFinder<AccountId: Parameter> {
-    fn contracts(operator_id: &AccountId) -> Vec<AccountId>;
-}
-
-pub trait TransferOperator<AccountId: Parameter>: OperatorFinder<AccountId> {
-    /// Changes an operator for identified contracts with verify.
-    fn transfer_operator(
-        current_operator: AccountId,
-        contracts: Vec<AccountId>,
-        new_operator: AccountId,
-    ) -> Result<(), DispatchError> {
-        Self::verify_transfer_operator(&current_operator, &contracts)?;
-        Self::force_transfer_operator(current_operator, contracts, new_operator);
-        Ok(())
-    }
-
-    fn verify_transfer_operator(
-        current_operator: &AccountId,
-        contracts: &Vec<AccountId>,
-    ) -> Result<(), DispatchError> {
-        let operate_contracts = Self::contracts(current_operator);
-
-        // check the actually operate the contract.
-        if !contracts.iter().all(|c| operate_contracts.contains(c)) {
-            Err(DispatchError::Other(
-                "The sender don't operate the contracts address.",
-            ))?
-        }
-        Ok(())
-    }
-
-    /// Force Changes an operator for identified contracts without verify.
-    fn force_transfer_operator(
-        current_operator: AccountId,
-        contracts: Vec<AccountId>,
-        new_operator: AccountId,
-    );
-}
+use pallet_plasm_support::{ContractFinder, OperatorFinder, TransferOperator};
 
 /// The module's configuration trait.
 pub trait Trait: pallet_contracts::Trait {
