+++ frame/plasm-support/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "pallet-plasm-support"
+version = "1.7.0"
+authors = ["Stake Technologies <devops@stake.co.jp>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://docs.plasmnet.io/"
+repository = "https://github.com/staketechnologies/Plasm/"
+description = "FRAME support pallet for Plasm traits"
+
+[dependencies]
+sp-runtime = { version = "2.0.0", default_features = false }
+frame-support = { version = "2.0.0", default_features = false }
+sp-std = { version = "2.0.0", default_features = false }
+
+
+[dev-dependencies]
+
+
+[features]
+default = ["std"]
+std = [
+    "sp-runtime/std",
+    "frame-support/std",
+    "sp-std/std",
+]
