+++ frame/plasm-validator/Cargo.toml
@@ -22,7 +22,7 @@ frame-support = { version = "2.0.0", default_features = false }
 frame-system = { version = "2.0.0", default_features = false }
 sp-staking = { version = "2.0.0", default_features = false }
 
-pallet-plasm-rewards = { path = "../plasm-rewards", version = "1.4.0", default_features = false }
+pallet-plasm-rewards = { path = "../plasm-rewards", version = "1.7.0", default_features = false }
 
 [dev-dependencies]
 hex-literal = "0.2.1"
