+++ frame/trading/Cargo.toml
@@ -20,7 +20,7 @@ frame-support = { version = "2.0.0", default_features = false }
 pallet-timestamp = { version = "2.0.0", default_features = false }
 frame-system = { version = "2.0.0", default_features = false }
 
-pallet-contract-operator = { path = "../operator", version = "1.4.0", default_features = false }
+pallet-plasm-support = { path = "../plasm-support", version = "1.7.0", default_features = false }
 
 [features]
 default = ["std"]
@@ -35,5 +35,5 @@ std = [
     "frame-support/std",
     "pallet-timestamp/std",
     "frame-system/std",
-    "pallet-contract-operator/std",
+    "pallet-plasm-support/std",
 ]
